Forward-merge branch-23.06 to branch-23.08

+++ cpp/include/kvikio/utils.hpp
@@ -39,7 +39,7 @@ inline constexpr std::size_t page_size = 4096;
   return static_cast<off_t>(x);
 }
 
-[[nodiscard]] inline off_t convert_size2ssize(std::size_t x)
+[[nodiscard]] inline ssize_t convert_size2ssize(std::size_t x)
 {
   if (x >= static_cast<std::size_t>(std::numeric_limits<ssize_t>::max())) {
     throw CUfileException("size_t argument too large to fit ssize_t");
